+++ spec/rubocop/cop/lint/erb_new_arguments_spec.rb
@@ -19,6 +19,10 @@ RSpec.describe RuboCop::Cop::Lint::ErbNewArguments, :config do
         ERB.new(str, nil)
                      ^^^ Passing safe_level with the 2nd argument of `ERB.new` is deprecated. Do not use it, and specify other arguments as keyword arguments.
       RUBY
+
+      expect_correction(<<~RUBY)
+        ERB.new(str)
+      RUBY
     end
 
     it 'registers an offense when using `ERB.new` ' \
@@ -27,6 +31,10 @@ RSpec.describe RuboCop::Cop::Lint::ErbNewArguments, :config do
         ERB.new(str, nil, '-')
                           ^^^ Passing trim_mode with the 3rd argument of `ERB.new` is deprecated. Use keyword argument like `ERB.new(str, trim_mode: '-')` instead.
                      ^^^ Passing safe_level with the 2nd argument of `ERB.new` is deprecated. Do not use it, and specify other arguments as keyword arguments.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        ERB.new(str, trim_mode: '-')
       RUBY
     end
 
@@ -37,6 +45,10 @@ RSpec.describe RuboCop::Cop::Lint::ErbNewArguments, :config do
                                ^^^^^^^^^^^^^^^^ Passing eoutvar with the 4th argument of `ERB.new` is deprecated. Use keyword argument like `ERB.new(str, eoutvar: '@output_buffer')` instead.
                           ^^^ Passing trim_mode with the 3rd argument of `ERB.new` is deprecated. Use keyword argument like `ERB.new(str, trim_mode: '-')` instead.
                      ^^^ Passing safe_level with the 2nd argument of `ERB.new` is deprecated. Do not use it, and specify other arguments as keyword arguments.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        ERB.new(str, trim_mode: '-', eoutvar: '@output_buffer')
       RUBY
     end
 
@@ -49,6 +61,10 @@ RSpec.describe RuboCop::Cop::Lint::ErbNewArguments, :config do
                           ^^^ Passing trim_mode with the 3rd argument of `ERB.new` is deprecated. Use keyword argument like `ERB.new(str, trim_mode: '-')` instead.
                      ^^^ Passing safe_level with the 2nd argument of `ERB.new` is deprecated. Do not use it, and specify other arguments as keyword arguments.
       RUBY
+
+      expect_correction(<<~RUBY)
+        ERB.new(str, trim_mode: '-', eoutvar: '@output_buffer')
+      RUBY
     end
 
     it 'registers an offense when using `ERB.new` ' \
@@ -59,6 +75,10 @@ RSpec.describe RuboCop::Cop::Lint::ErbNewArguments, :config do
                           ^^^ Passing trim_mode with the 3rd argument of `ERB.new` is deprecated. Use keyword argument like `ERB.new(str, trim_mode: '-')` instead.
                      ^^^ Passing safe_level with the 2nd argument of `ERB.new` is deprecated. Do not use it, and specify other arguments as keyword arguments.
       RUBY
+
+      expect_correction(<<~RUBY)
+        ERB.new(str, trim_mode: '-', eoutvar: '@output_buffer')
+      RUBY
     end
 
     it 'registers an offense when using `::ERB.new` ' \
@@ -68,6 +88,10 @@ RSpec.describe RuboCop::Cop::Lint::ErbNewArguments, :config do
                                  ^^^^^^^^^^^^^^^^ Passing eoutvar with the 4th argument of `ERB.new` is deprecated. Use keyword argument like `ERB.new(str, eoutvar: '@output_buffer')` instead.
                             ^^^ Passing trim_mode with the 3rd argument of `ERB.new` is deprecated. Use keyword argument like `ERB.new(str, trim_mode: '-')` instead.
                        ^^^ Passing safe_level with the 2nd argument of `ERB.new` is deprecated. Do not use it, and specify other arguments as keyword arguments.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        ::ERB.new(str, trim_mode: '-', eoutvar: '@output_buffer')
       RUBY
     end